Added some basics of connection processing; not much but believe me when I say that I wrote about 50x that amount of code and then rewrote and rewrote into oblivion (protocol & sockettype adjustments).
I also added a JSON library, which to be honest I'm not sure I'll use. I was planning on using it for every command, but because JSON cannot be processed until fully received, it cannot be used for example for transferring the filelist (I want to start processing files before the entire filelist has arrived).