Added try-catch to Source to prevent in-use crash
This crash would occur eg on Windows when backing up a program that is
currently executing. canRead() would return true but later on things would go
to shit. Later on this will obviously include a hook to let the UI know some
file was not backed up. The UI can then show eg "warning" instead of "ok"
status for that backup and display extra info when necessary.