Fixed bug that causes client not to exit
bacsy-client should exit after running a single source once. It did that, but
when a second source is present with multiple timers, and that source is
disabled, bacsy-client no longer exits. This was fixed by simply checking in
the Source::createTimers() function if this source is enabled and if not
exiting. It might be preferrable to not even put Sources that are disabled in
the Sources list. I don't know. Both approaches have their advantages and
disadvantages.