Add file record keeping of last run
An interface is defined and both plain text writer and reader classes are
defined. This is also tested reasonably well. The whole thing with the
separate write and read implementations was necessary to avoid static factory
methods because both write and read constructors need a filename. Only the
write filename is to write to and the read filename is to read from. If the
filename is read from in the write record keeper, the in memory database is
pre-populated, which is not something we'd like (clear() functions would be
necessary to erase the in memory representation).